refactor(payout-service): refactor code of tests
